Fix core engine program modeler errors
