+++ debug/modelPrograms/sanity-snapshot/output-file-digest-list.yml
@@ -19,28 +19,28 @@
   - "fileName": "test-cached-on-disk-datum"
     "digest": "1238a49727662ab83b402a00df2c436977491023"
   - "fileName": "test-get-custom-typed-datum"
-    "digest": "c8056a9b791fb4c9144a3c9ca47500eb0f0d8dea"
+    "digest": "36602052ff7275309656c218d218c516c388e88f"
   - "fileName": "test-get-type-script-typed-datum"
-    "digest": "b70098467bc701104c971d6b28e8a72979402c2a"
+    "digest": "ae33b23dfb4b50d684fe1ec2cac4d810e992dd39"
   - "fileName": "test-graph-render"
     "digest": "ded88f4516c06bf398ab19a646417a89000f4c4b"
   - "fileName": "test-joining-one-to-many"
-    "digest": "826ac8e2c63ebc9059de8c24ead96910d4d5baee"
+    "digest": "39e8943a64c97f48c1bfef1b5a3ee3c21f25c8fe"
   - "fileName": "test-joining-one-to-one"
-    "digest": "ce2b7fe9bf6bc37bbf4496525c31a4d8305f1542"
+    "digest": "54e8307583992cbc3aef9c43aa2fb516b217eb95"
   - "fileName": "test-joining-one-to-voictent"
     "digest": "3b7cd4f7d222991b2d00a3f97be6367d94a17d31"
   - "fileName": "test-joining-voictent-to-voictent"
-    "digest": "0e75faa5305abdcd6f18970371aecfc1b361fe3b"
+    "digest": "bbdf1bea94ef141d43edf2861a34b2e601a07c39"
   - "fileName": "test-json-serialization"
-    "digest": "e42576f32504918bde63398cc0d9a632c606f781"
+    "digest": "e2cfbff643dd0ff6309b851601b47fce511bafd7"
   - "fileName": "test-releasing-a-left-voictent-multiple-times"
-    "digest": "e3431de9ddfec788fa684d8c2d102c18bae38b41"
+    "digest": "aed5522375efb42384f51ac79d7c975182d3ed10"
   - "fileName": "test-releasing-a-right-voictent-multiple-times"
-    "digest": "0ecd8a1c8b356d466b6f5304f8714f0a92986c51"
+    "digest": "c8bc29eba7a7fb407d4e7a3263222eed9ccadc13"
   - "fileName": "test-serialize"
-    "digest": "284166ee9ec68ee1fce51dc11c20ddcfe37a0abc"
+    "digest": "586f98ff5ce0dad302ad0194e2df136c38edec38"
   - "fileName": "test-typed-datum"
     "digest": "6c9e5d42c99a80192547948b715d910add3a5fd3"
   - "fileName": "test-voictent-input"
-    "digest": "1f7a76f59edbeca9a780b3c537d40650c7371a21"
+    "digest": "8f78bbc5bf5303794cebfcc603f4e60bf962518d"
