+++ packages/voictents-and-estinants-engine/src/example-programs/engine-behavior/testJoiningOneToMany.ts
@@ -58,6 +58,10 @@ const joinCollections: Estinant2<
   },
 };
 
+/**
+ * Tests a transform that consumes multiple items from a right collection for
+ * each item of the left collection
+ */
 digikikify2({
   inputVoictentList: [
     new InMemoryVoictent<Voque1>({
