--- packages/voictents-and-estinants-engine/src/example-programs/getTypedTestCaseInputTypeName.ts
@@ -1,46 +1,0 @@
-import { Estinant2 } from '../core/engine-shell/estinant/estinant';
-import { LeftInputHubblepupVicken } from '../core/engine-shell/vicken/leftInputVicken';
-import { OutputVicken } from '../core/engine-shell/vicken/outputVicken';
-import {
-  DATUM_TEST_CASE_INPUT_GEPP,
-  DatumTestCaseInputVoque,
-} from '../custom/programmable-units/datum-test-case-input/datumTestCaseInput';
-import { getTypeScriptTypedDatum } from '../utilities/typed-datum/type-script/typeScriptTypedDatum';
-import {
-  SERIALIZABLE_TYPE_NAME_GEPP,
-  SerializableTypeNameVoque,
-} from './serializableTypeName';
-
-export const getTypedTestCaseInputTypeName: Estinant2<
-  LeftInputHubblepupVicken<DatumTestCaseInputVoque>,
-  [],
-  OutputVicken<[SerializableTypeNameVoque]>
-> = {
-  version: 2,
-  name: 'getTypedTestCaseInputTypeName',
-  leftInputAppreffinge: {
-    gepp: DATUM_TEST_CASE_INPUT_GEPP,
-    isWibiz: false,
-  },
-  rightInputAppreffingeTuple: [],
-  outputAppreffinge: {
-    geppTuple: [SERIALIZABLE_TYPE_NAME_GEPP],
-  },
-  tropoig: (input) => {
-    const inputOdeshin = input.hubblepup;
-    const testCaseInput = inputOdeshin.grition;
-
-    const typedDatum = getTypeScriptTypedDatum(testCaseInput);
-
-    const output = {
-      zorn: inputOdeshin.zorn,
-      grition: {
-        typeName: typedDatum.typeName,
-      },
-    };
-
-    return {
-      [SERIALIZABLE_TYPE_NAME_GEPP]: [output],
-    };
-  },
-};
