+++ packages/voictents-and-estinants-engine/src/example-programs/testGetCustomTypedDatum.ts
@@ -1,19 +1,61 @@
+import { Estinant2 } from '../core/engine-shell/estinant/estinant';
+import { LeftInputHubblepupVicken } from '../core/engine-shell/vicken/leftInputVicken';
+import { OutputVicken } from '../core/engine-shell/vicken/outputVicken';
 import { digikikify2 } from '../core/engine/digikikify';
 import { InMemoryOdeshinVoictent } from '../core/engine/inMemoryOdeshinVoictent';
 import { InMemoryVoque } from '../core/engine/inMemoryVoictent';
 import {
   DATUM_TEST_CASE_INPUT_GEPP,
   DATUM_TEST_CASE_INPUT_ODESHIN_LIST,
+  DatumTestCaseInputVoque,
 } from '../custom/programmable-units/datum-test-case-input/datumTestCaseInput';
+import { getCustomTypedDatum } from '../utilities/typed-datum/customTypedDatum';
 import { buildAddMetadataForSerialization } from './buildAddMetadataForSerialization';
-import { getCustomTypedTestCaseInputTypeName } from './getCustomTypedTestCaseInputTypeName';
 import {
   JsonSerializableVoictent,
   JsonSerializableVoque,
 } from './jsonSerializableVoictent';
+import {
+  SerializableTypeNameVoque,
+  SERIALIZABLE_TYPE_NAME_GEPP,
+} from './serializableTypeName';
 
 type SerializedConfiguration = JsonSerializableVoque<'serialized'>;
 type TypedDatumVoque = InMemoryVoque<'typed-datum', unknown>;
+
+const getCustomTypedTestCaseInputTypeName: Estinant2<
+  LeftInputHubblepupVicken<DatumTestCaseInputVoque>,
+  [],
+  OutputVicken<[SerializableTypeNameVoque]>
+> = {
+  version: 2,
+  name: 'getCustomTypedTestCaseInputTypeName',
+  leftInputAppreffinge: {
+    gepp: DATUM_TEST_CASE_INPUT_GEPP,
+    isWibiz: false,
+  },
+  rightInputAppreffingeTuple: [],
+  outputAppreffinge: {
+    geppTuple: [SERIALIZABLE_TYPE_NAME_GEPP],
+  },
+  tropoig: (input) => {
+    const inputOdeshin = input.hubblepup;
+    const testCaseInput = inputOdeshin.grition;
+
+    const typedDatum = getCustomTypedDatum(testCaseInput);
+
+    const output = {
+      zorn: inputOdeshin.zorn,
+      grition: {
+        typeName: typedDatum.typeName,
+      },
+    };
+
+    return {
+      [SERIALIZABLE_TYPE_NAME_GEPP]: [output],
+    };
+  },
+};
 
 /**
  * Tests the "getCustomTypedDatum" function. Commit the output to get a signal
