+++ packages/voictents-and-estinants-engine/src/example-programs/testGetTypeScriptTypedDatum.ts
@@ -1,11 +1,15 @@
+import { Estinant2 } from '../core/engine-shell/estinant/estinant';
+import { LeftInputHubblepupVicken } from '../core/engine-shell/vicken/leftInputVicken';
+import { OutputVicken } from '../core/engine-shell/vicken/outputVicken';
 import { digikikify2 } from '../core/engine/digikikify';
 import { InMemoryOdeshinVoictent } from '../core/engine/inMemoryOdeshinVoictent';
 import {
   DATUM_TEST_CASE_INPUT_GEPP,
   DATUM_TEST_CASE_INPUT_ODESHIN_LIST,
+  DatumTestCaseInputVoque,
 } from '../custom/programmable-units/datum-test-case-input/datumTestCaseInput';
+import { getTypeScriptTypedDatum } from '../utilities/typed-datum/type-script/typeScriptTypedDatum';
 import { buildAddMetadataForSerialization } from './buildAddMetadataForSerialization';
-import { getTypedTestCaseInputTypeName } from './getTypedTestCaseInputTypeName';
 import {
   JsonSerializableVoictent,
   JsonSerializableVoque,
@@ -16,6 +20,40 @@ import {
 } from './serializableTypeName';
 
 type SerializedConfiguration = JsonSerializableVoque<'serialized'>;
+
+export const getTypedTestCaseInputTypeName: Estinant2<
+  LeftInputHubblepupVicken<DatumTestCaseInputVoque>,
+  [],
+  OutputVicken<[SerializableTypeNameVoque]>
+> = {
+  version: 2,
+  name: 'getTypedTestCaseInputTypeName',
+  leftInputAppreffinge: {
+    gepp: DATUM_TEST_CASE_INPUT_GEPP,
+    isWibiz: false,
+  },
+  rightInputAppreffingeTuple: [],
+  outputAppreffinge: {
+    geppTuple: [SERIALIZABLE_TYPE_NAME_GEPP],
+  },
+  tropoig: (input) => {
+    const inputOdeshin = input.hubblepup;
+    const testCaseInput = inputOdeshin.grition;
+
+    const typedDatum = getTypeScriptTypedDatum(testCaseInput);
+
+    const output = {
+      zorn: inputOdeshin.zorn,
+      grition: {
+        typeName: typedDatum.typeName,
+      },
+    };
+
+    return {
+      [SERIALIZABLE_TYPE_NAME_GEPP]: [output],
+    };
+  },
+};
 
 /**
  * Tests the "getTypeScriptTypedDatum" function. Commit the output to keep track
