+++ packages/voictents-and-estinants-engine/src/example-programs/testJsonSerialization.ts
@@ -7,6 +7,12 @@ import {
 
 type SerializedConfiguration = JsonSerializableVoque<'serialized'>;
 
+/**
+ * Tests the "JsonSerializableVoictent" by initializing it with data.
+ *
+ * @note Running the collection through the engine is technically unnecessary,
+ * but it allows the program modeler to construct a snapshot digest of this file
+ */
 digikikify2({
   inputVoictentList: [
     new JsonSerializableVoictent<SerializedConfiguration>({
