+++ packages/voictents-and-estinants-engine/src/example-programs/testSerialize.ts
@@ -5,6 +5,12 @@ import { SerializableVoictent } from './serializableVoictent';
 
 type SerializedConfiguration = JsonSerializableVoque<'serialized'>;
 
+/**
+ * Tests the "SerializableVoictent" by initializing it with data.
+ *
+ * @note Running the collection through the engine is technically unnecessary,
+ * but it allows the program modeler to construct a snapshot digest of this file
+ */
 digikikify2({
   inputVoictentList: [
     // eslint-disable-next-line no-new
